requirements.txt
pycryptodome==3.17
pywin32==305
tqdm==4.66.1
